+++ RPGatorDll/CurrentMapObject.cpp
@@ -28,7 +28,7 @@ CurrentGameObject<T>::CurrentGameObject(sqliteResult, mapObjects, mapObjectsCoun
 	currentLocation = location;
 	if (base->mesh)
 	{
-		node = Render::instance->createNode(false, base->mesh, base->texture, false, vector3df(base->scale / 10.0f, base->scale / 10.0f, base->scale / 10.0f), vector3df((f32)x * CELL_SIZE, 0.0f, (f32)y * CELL_SIZE));
+		node = Render::instance->createNode(false, base->mesh, base->texture, false, vector3df(base->scale / 10.0f, base->scale / 10.0f, base->scale / 10.0f), vector3df(x * CELL_SIZE, 0.0f, y * CELL_SIZE));
 	}
 	else
 		node = NULL;
@@ -43,7 +43,7 @@ CurrentGameObject<T>::CurrentGameObject(currentMapObjectSpawnedPacket, mapObject
 	currentLocation = NULL; //Client does not use it. It uses Universe->currentLocation
 	if (base->mesh)
 	{
-		node = Render::instance->createNode(false, base->mesh, base->texture, false, vector3df(base->scale / 10.0f, base->scale / 10.0f, base->scale / 10.0f), vector3df((f32)x * CELL_SIZE, 0.0f, (f32)y * CELL_SIZE));
+		node = Render::instance->createNode(false, base->mesh, base->texture, false, vector3df(base->scale / 10.0f, base->scale / 10.0f, base->scale / 10.0f), vector3df(x * CELL_SIZE, 0.0f, y * CELL_SIZE));
 	}
 	else
 		node = NULL;
